requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
oauthlib==3.2.0
phonenumbers==8.12.44
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.9
zipp==3.7.0
